# iac/files/openapi_with_rate_limits.yaml
openapi: 3.0.0
info:
  title: AI Content Factory API
  description: Production-grade API for AI-powered content generation with enterprise rate limiting
  version: "1.0.0"
  contact:
    name: AI Content Factory Team
    email: support@ai-content-factory.com

servers:
  - url: ${backend_url}
    description: Production backend server

# Rate limiting configuration
x-google-ratelimit:
  quotas:
    - name: "requests-per-minute-per-ip"
      limit: 10
      interval: "1m"
      region: "global"
      key: "$remote_addr"
    - name: "requests-per-hour-per-api-key"
      limit: 100
      interval: "1h"
      region: "global"
      key: "$http_x_api_key"
    - name: "expensive-operations-per-day"
      limit: 50
      interval: "1d"
      region: "global"
      key: "$http_x_api_key"
      applies_to:
        - "/api/v1/content/generate"

paths:
  /api/v1/content/generate:
    post:
      summary: Generate comprehensive educational content
      description: Creates content outline and derivative materials with rate limiting
      operationId: generateContent
      security:
        - ApiKeyAuth: []
      x-google-quota:
        metricCosts:
          "requests-per-minute-per-ip": 1
          "requests-per-hour-per-api-key": 1
          "expensive-operations-per-day": 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                syllabus_text:
                  type: string
                  minLength: 50
                  maxLength: 5000
                  description: Educational content to process
                target_formats:
                  type: array
                  items:
                    type: string
                    enum: ["podcast_script", "study_guide", "faq_collection", "flashcard_collection", "one_pager_summary", "detailed_reading_material", "reading_guide_questions"]
                  description: Requested content formats
              required:
                - syllabus_text
      responses:
        '200':
          description: Content generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  content_outline:
                    type: object
                  generated_content:
                    type: object
        '400':
          description: Invalid request parameters
        '401':
          description: Authentication required
        '422':
          description: Validation error
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Rate limit exceeded. Try again in 60 seconds."
                  retry_after:
                    type: integer
                    example: 60

  /api/v1/jobs:
    get:
      summary: List content generation jobs
      operationId: listJobs
      security:
        - ApiKeyAuth: []
      x-google-quota:
        metricCosts:
          "requests-per-minute-per-ip": 1
          "requests-per-hour-per-api-key": 1
      responses:
        '200':
          description: Jobs retrieved successfully
        '401':
          description: Authentication required
        '429':
          description: Rate limit exceeded

  /api/v1/jobs/{job_id}:
    get:
      summary: Get job status and results
      operationId: getJob
      security:
        - ApiKeyAuth: []
      x-google-quota:
        metricCosts:
          "requests-per-minute-per-ip": 1
          "requests-per-hour-per-api-key": 1
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status retrieved
        '404':
          description: Job not found
        '429':
          description: Rate limit exceeded

  /health:
    get:
      summary: Health check endpoint
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy

  /api/v1/auth/token:
    post:
      summary: Authenticate and get access token
      operationId: authenticate
      x-google-quota:
        metricCosts:
          "requests-per-minute-per-ip": 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                api_key:
                  type: string
                  description: Client API key
              required:
                - api_key
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: "bearer"
                  expires_in:
                    type: integer
                    example: 3600
        '401':
          description: Invalid credentials
        '429':
          description: Rate limit exceeded

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        trace_id:
          type: string
          description: Request trace ID
      required:
        - error

    RateLimitError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            retry_after:
              type: integer
              description: Seconds to wait before retrying
            quota_used:
              type: integer
              description: Current quota usage
            quota_limit:
              type: integer
              description: Quota limit
          required:
            - retry_after

# Global rate limiting rules
x-google-endpoints:
  - name: "ai-content-factory.endpoints.PROJECT_ID.cloud.goog"
    allowCors: true

x-google-management:
  metrics:
    - name: "requests-per-minute-per-ip"
      displayName: "Requests per minute per IP"
      description: "Rate limiting metric for requests per minute per IP address"
      metricKind: "GAUGE"
      valueType: "INT64"
    - name: "requests-per-hour-per-api-key"
      displayName: "Requests per hour per API key"
      description: "Rate limiting metric for requests per hour per API key"
      metricKind: "GAUGE"
      valueType: "INT64"
    - name: "expensive-operations-per-day"
      displayName: "Expensive operations per day"
      description: "Daily quota for resource-intensive operations"
      metricKind: "GAUGE"
      valueType: "INT64"

  quota:
    limits:
      - name: "requests-per-minute-per-ip"
        metric: "requests-per-minute-per-ip"
        unit: "1/min/{ip}"
        values:
          STANDARD: 10
      - name: "requests-per-hour-per-api-key"
        metric: "requests-per-hour-per-api-key"
        unit: "1/h/{api_key}"
        values:
          STANDARD: 100
      - name: "expensive-operations-per-day"
        metric: "expensive-operations-per-day"
        unit: "1/d/{api_key}"
        values:
          STANDARD: 50